# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Remedy Scribe API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/notes/{note_id}": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Get Note",
        "operationId": "get_note_api_v1_notes__note_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "notes"
        ],
        "summary": "Edit Section",
        "description": "P0-5: \"Doctor can freely edit any section before signing; edits are\ntracked for the edit-burden metric.\" Every edit writes a NoteRevision\nregardless of size \u2014 the edit-burden metric needs the raw edit\nhistory, not just the final diff from generation.",
        "operationId": "edit_section_api_v1_notes__note_id__patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteSectionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}/grounding": {
      "get": {
        "tags": [
          "notes"
        ],
        "summary": "Read Grounding",
        "description": "Phase 3 (P0-7): everything needed to answer \"where did this line come\nfrom?\" in one read.\n\nResolves each section's stored spans against the note's *current* text,\nreturns the cited transcript passages with their audio timestamps, and\nreports which rung of the degradation ladder this encounter is on\n(audio + transcript, transcript only, or neither). See\napp/services/grounding.py for why each of those is verified rather than\nassumed.\n\nAudited separately from `note.read`: reading a note is reading the\nclinician-facing summary, while this returns verbatim transcript\npassages \u2014 a strictly larger PHI disclosure, and one worth being able\nto account for on its own.",
        "operationId": "read_grounding_api_v1_notes__note_id__grounding_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroundingOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/notes/{note_id}/transition": {
      "post": {
        "tags": [
          "notes"
        ],
        "summary": "Transition Note",
        "description": "Drives the P0-5 state machine one step at a time. Signing\n(to_status == \"signed\") additionally requires prc_license_number and\nis recorded in the audit trail with the clinician's identity.",
        "operationId": "transition_note_api_v1_notes__note_id__transition_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "note_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Note Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioState": {
        "type": "string",
        "enum": [
          "available",
          "never_recorded",
          "withdrawn",
          "expired",
          "unreachable"
        ],
        "title": "AudioState",
        "description": "The degradation ladder the Phase 3 heads-up asks for: notes outlive\naudio, and the doctor must understand *which* state they are in rather\nthan seeing a play button that does nothing."
      },
...
      "GroundedSectionOut": {
        "properties": {
          "suppressed": {
            "type": "boolean",
            "title": "Suppressed"
          },
          "spans": {
            "items": {
              "$ref": "#/components/schemas/GroundedSpanOut"
            },
            "type": "array",
            "title": "Spans"
          },
          "spans_fit": {
            "type": "boolean",
            "title": "Spans Fit"
          },
          "edited_since_generation": {
            "type": "boolean",
            "title": "Edited Since Generation"
          }
        },
        "type": "object",
        "required": [
          "suppressed",
          "spans",
          "spans_fit",
          "edited_since_generation"
        ],
        "title": "GroundedSectionOut"
      },
      "GroundedSegmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "speaker": {
            "type": "string",
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "start_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Ms"
          },
          "end_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Ms"
          },
          "cited": {
            "type": "boolean",
            "title": "Cited"
          }
        },
        "type": "object",
        "required": [
          "id",
          "index",
          "speaker",
          "text",
          "start_ms",
          "end_ms",
          "cited"
        ],
        "title": "GroundedSegmentOut",
        "description": "One transcript passage. `start_ms`/`end_ms` are nullable because a\npersisted segment with no words is representable, and a read endpoint\nshould degrade to \"this passage cannot be played\" rather than 500."
      },
      "GroundedSpanOut": {
        "properties": {
          "text_start": {
            "type": "integer",
            "title": "Text Start"
          },
          "text_end": {
            "type": "integer",
            "title": "Text End"
          },
          "segment_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Segment Ids"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text_start",
          "text_end",
          "segment_ids",
          "text"
        ],
        "title": "GroundedSpanOut"
      },
      "GroundingOut": {
        "properties": {
          "note_id": {
            "type": "string",
            "title": "Note Id"
          },
          "encounter_id": {
            "type": "string",
            "title": "Encounter Id"
          },
          "audio_state": {
            "$ref": "#/components/schemas/AudioState"
          },
          "transcript_state": {
            "$ref": "#/components/schemas/TranscriptState"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/GroundedSegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "sections": {
            "additionalProperties": {
              "$ref": "#/components/schemas/GroundedSectionOut"
            },
            "type": "object",
            "title": "Sections"
          }
        },
        "type": "object",
        "required": [
          "note_id",
          "encounter_id",
          "audio_state",
          "transcript_state",
          "segments",
          "sections"
        ],
        "title": "GroundingOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NoteOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "encounter_id": {
            "type": "string",
            "title": "Encounter Id"
          },
          "status": {
            "$ref": "#/components/schemas/NoteStatus"
          },
          "assessment": {
            "type": "string",
            "title": "Assessment"
          },
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "subjective": {
            "type": "string",
            "title": "Subjective"
          },
          "objective": {
            "type": "string",
            "title": "Objective"
          },
          "note_generator_provider": {
            "type": "string",
            "title": "Note Generator Provider"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "signed_by_clinician_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed By Clinician Id"
          },
          "signed_prc_license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed Prc License Number"
          },
          "signed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "encounter_id",
          "status",
          "assessment",
          "plan",
          "subjective",
          "objective",
          "note_generator_provider",
          "prompt_version",
          "signed_by_clinician_id",
          "signed_prc_license_number",
          "signed_at"
        ],
        "title": "NoteOut"
      },
      "NoteSectionUpdate": {
        "properties": {
          "section": {
            "type": "string",
            "enum": [
              "assessment",
              "plan",
              "subjective",
              "objective"
            ],
            "title": "Section"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "section",
          "text"
        ],
        "title": "NoteSectionUpdate",
        "description": "Doctor edit to one section before signing (P0-5). Recorded as a\nNoteRevision for the edit-burden metric regardless of how small."
      },
      "NoteStatus": {
        "type": "string",
        "enum": [
          "generated",
          "filed",
          "authenticated",
          "signed"
        ],
        "title": "NoteStatus",
        "description": "P0-5: \"Note state machine enforces four distinct states: generated\n\u2192 filed \u2192 authenticated \u2192 signed, with no state skippable.\"\n\nThe Enum column type below constrains the DB to these four values \u2014\nliterally, via a CHECK constraint (`create_constraint=True`; see\ndocs/decisions/0010 for why that flag has to be explicit: SQLAlchemy\n2.0 does not add the constraint by default for a non-native enum,\nwhich meant this docstring's claim was false until Phase 0.4). The\n*ordering* (no skipping) is enforced separately, in\napp/services/note_lifecycle.py, which is the only code path allowed\nto write Note.status \u2014 routes must go through it rather than setting\nstatus directly."
      },
      "NoteTransitionRequest": {
        "properties": {
          "to_status": {
            "$ref": "#/components/schemas/NoteStatus"
          },
          "prc_license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prc License Number"
          },
          "confirmed_patient_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed Patient Id"
          }
        },
        "type": "object",
        "required": [
          "to_status"
        ],
        "title": "NoteTransitionRequest",
        "description": "Advances the note exactly one step in the state machine\n(P0-5: generated -> filed -> authenticated -> signed, no skipping).\nSigning additionally requires PRC license number, and **filing\nadditionally requires confirming the patient** (P0-6: identity is\nre-confirmed at the moment a note is filed, not only at recording\nstart)."
      },
...
      "TranscriptState": {
        "type": "string",
        "enum": [
          "available",
          "never_transcribed",
          "withdrawn",
          "expired"
        ],
        "title": "TranscriptState",
        "description": "The same ladder as AudioState, and it needs the same rung for the\nsame reason.\n\nPhase 4.4 added a retention job that deletes a withdrawn encounter's\ntranscript, not just its audio \u2014 correctly, since the transcript is\nverbatim PHI derived from a recording the patient asked to have\ndestroyed. But with only EXPIRED available, that deletion was\ndescribed to the doctor as \"the retention period elapsed\", which is\nthe wrong reason. Decision 0030's whole argument for a five-state\naudio ladder was that a withdrawal and a retention expiry are\nobservably identical and mean entirely different things; a\nthree-state transcript ladder quietly contradicted it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}